In-and-outbot/requirements.txt
ccxt
pandas
SQLAlchemy
python-dotenv
